bugfix: using the write project instead of the compute one
When using `pandas_gbq.read_gbq`, the incorrect project ID was passed (write instead of compute).

Since this method doesn't support passing a separate compute project, we also need to correctly specify the reference to the dataset's project in the query, hence the changes in `list_tables` and `list_columns`.