feat: add option to specify BQ storage billing mode (#95)
BigQuery provides 2 different billing modes for storage: `LOGICAL` or
`PHYSICAL`.

Depending on the data, one of the model may be cheaper.

This option enables to run `lea` with a `--storage-billing-mode` that is
used when creating the dataset to set the billing mode.

NB: the billing mode will only be applied when the dataset is created,
so if it already exists the billing mode will not be changed by Lea and
must be changed by directly editing the dataset (this may be done via a
simple SQL query).

---------

Co-authored-by: Max Halford <maxhalford25@gmail.com>