update all deps within their authorized versions (#122)
Due to a security vulnerability:

![image](https://github.com/user-attachments/assets/b7f1bb44-0d6e-45e2-a54f-d936be59b20a)

Only update poetry.lock, so this has no effect when lea is installed as
a lib. There is no need for a release.
    
<!-- This is an auto-generated description by cubic. -->
---

## Summary by cubic
Updated all dependencies in poetry.lock to their latest authorized
versions to address a security vulnerability. This change only affects
the lockfile and does not impact users installing the package as a
library.

<!-- End of auto-generated description by cubic. -->